Rename _metadict to AssocList and move it to theano.gof.utils
